refactor: review requests
- move suggested fees service into nearer located file
- remove if-clause to terminate suggested relayer fee population if already exists
